move hive parser to a separate module
