+++ flink-table/flink-sql-parser-hive/src/main/codegen/data/Parser.tdd
@@ -16,18 +16,18 @@
 
 {
   # Generated parser implementation package and class name.
-  package: "org.apache.flink.sql.parser.impl",
+  package: "org.apache.flink.sql.parser.hive.impl",
   class: "FlinkHiveSqlParserImpl",
 
   # List of additional classes and packages to import.
   # Example. "org.apache.calcite.sql.*", "java.util.List".
   # Please keep the import classes in alphabetical order if new class is added.
   imports: [
-    "org.apache.flink.sql.parser.ddl.hive.HiveDDLUtils"
-    "org.apache.flink.sql.parser.ddl.hive.SqlAlterHiveDatabaseLocation"
-    "org.apache.flink.sql.parser.ddl.hive.SqlAlterHiveDatabaseOwner"
-    "org.apache.flink.sql.parser.ddl.hive.SqlAlterHiveDatabaseProps"
-    "org.apache.flink.sql.parser.ddl.hive.SqlCreateHiveDatabase"
+    "org.apache.flink.sql.parser.hive.ddl.HiveDDLUtils"
+    "org.apache.flink.sql.parser.hive.ddl.SqlAlterHiveDatabaseLocation"
+    "org.apache.flink.sql.parser.hive.ddl.SqlAlterHiveDatabaseOwner"
+    "org.apache.flink.sql.parser.hive.ddl.SqlAlterHiveDatabaseProps"
+    "org.apache.flink.sql.parser.hive.ddl.SqlCreateHiveDatabase"
     "org.apache.flink.sql.parser.ddl.SqlAlterDatabase"
     "org.apache.flink.sql.parser.ddl.SqlAlterTable"
     "org.apache.flink.sql.parser.ddl.SqlAlterTableProperties"
