+++ flink-table/flink-sql-parser-hive/src/main/codegen/includes/compoundIdentifier.ftl
