+++ flink-table/flink-sql-parser-hive/src/main/codegen/includes/parserImpls.ftl
