+++ flink-table/flink-sql-parser-hive/src/main/java/org/apache/flink/sql/parser/hive/ddl/SqlAlterHiveDatabaseLocation.java
@@ -16,42 +16,33 @@
  * limitations under the License.
  */
 
-package org.apache.flink.sql.parser.ddl.hive;
+package org.apache.flink.sql.parser.hive.ddl;
 
 import org.apache.flink.sql.parser.ddl.SqlTableOption;
 
+import org.apache.calcite.sql.SqlCharStringLiteral;
 import org.apache.calcite.sql.SqlIdentifier;
 import org.apache.calcite.sql.SqlLiteral;
 import org.apache.calcite.sql.SqlNodeList;
 import org.apache.calcite.sql.parser.SqlParserPos;
 
+import static org.apache.flink.sql.parser.hive.ddl.SqlCreateHiveDatabase.DATABASE_LOCATION_URI;
+
 /**
- * ALTER Database owner DDL for Hive dialect.
+ * ALTER Database location DDL for Hive dialect.
  */
-public class SqlAlterHiveDatabaseOwner extends SqlAlterHiveDatabase {
+public class SqlAlterHiveDatabaseLocation extends SqlAlterHiveDatabase {
 
-	public static final String DATABASE_OWNER_NAME = "database.owner.name";
-	public static final String DATABASE_OWNER_TYPE = "database.owner.type";
-	public static final String USER_OWNER = "user";
-	public static final String ROLE_OWNER = "role";
-
-	public SqlAlterHiveDatabaseOwner(SqlParserPos pos, SqlIdentifier databaseName,
-			String ownerType, SqlIdentifier ownerName) {
+	public SqlAlterHiveDatabaseLocation(SqlParserPos pos, SqlIdentifier databaseName, SqlCharStringLiteral location) {
 		super(pos, databaseName, new SqlNodeList(pos));
-		SqlParserPos ownerPos = ownerName.getParserPosition();
 		getPropertyList().add(new SqlTableOption(
-				SqlLiteral.createCharString(DATABASE_OWNER_TYPE, ownerPos),
-				SqlLiteral.createCharString(ownerType, ownerPos),
-				ownerPos));
-		getPropertyList().add(new SqlTableOption(
-				SqlLiteral.createCharString(DATABASE_OWNER_NAME, ownerPos),
-				SqlLiteral.createCharString(ownerName.getSimple(), ownerPos),
-				ownerPos
-		));
+				SqlLiteral.createCharString(DATABASE_LOCATION_URI, location.getParserPosition()),
+				location,
+				location.getParserPosition()));
 	}
 
 	@Override
 	protected AlterHiveDatabaseOp getAlterOp() {
-		return AlterHiveDatabaseOp.CHANGE_OWNER;
+		return AlterHiveDatabaseOp.CHANGE_LOCATION;
 	}
 }
