+++ flink-table/pom.xml
@@ -45,6 +45,7 @@ under the License.
 		<module>flink-table-uber-blink</module>
 		<module>flink-sql-client</module>
 		<module>flink-sql-parser</module>
+		<module>flink-sql-parser-hive</module>
 	</modules>
 
 	<dependencyManagement>
